// src/components/doughnut/models/doughnut-data.ts
import { ChartData, Plugin } from "chart.js";

type AnyObject = Record<string, any>;

export interface DoughnutData {
  tooltipLabel?: number;
  value: number;
  unit: string;
  rate?: number;
}

export type DoughnutPlugin = Plugin<"doughnut", AnyObject>;

export type DoughnutChartData = ChartData<"doughnut", number[]>;